requirements.txt
google-generativeai
pyyaml
pandas
streamlit
langchain
PyPDF2
chromadb
faiss-cpu
langchain_google_genai
python-dotenv
langchain-community

